+++ internal/annotations/annotations.go
@@ -32,8 +32,9 @@ const (
 )
 
 const (
-	IngressClassKey        = "kubernetes.io/ingress.class"
-	KnativeIngressClassKey = "networking.knative.dev/ingress.class"
+	IngressClassKey                  = "kubernetes.io/ingress.class"
+	KnativeIngressClassKey           = "networking.knative.dev/ingress-class"
+	KnativeIngressClassDeprecatedKey = "networking.knative.dev/ingress.class"
 
 	AnnotationPrefix = "konghq.com"